import linux-2.4.x

+++ mmnommu/mlock.c
@@ -1,42 +1,35 @@
 /*
  *	linux/mm/mlock.c
  *
+ *  Copyright (c) 2001 Lineo, Inc. David McCullough <davidm@lineo.com>
+ *  Copyright (c) 2000-2001 D Jeff Dionne <jeff@uClinux.org> ref uClinux 2.0
  *  (C) Copyright 1995 Linus Torvalds
  */
-
-/*
- * uClinux revisions for NO_MM
- * Copyright (C) 1998  Kenneth Albanowski <kjahds@kjahds.com>,
- * Copyright (C) 1999,2000  D. Jeff Dionne <jeff@uclinux.org>,
- *                          Rt-Control, Inc./ Lineo Inc.
- */  
-
-#include <linux/stat.h>
-#include <linux/sched.h>
-#include <linux/kernel.h>
-#include <linux/mm.h>
+#include <linux/slab.h>
 #include <linux/shm.h>
-#include <linux/errno.h>
 #include <linux/mman.h>
-#include <linux/string.h>
-#include <linux/malloc.h>
+#include <linux/smp_lock.h>
+#include <linux/pagemap.h>
 
-asmlinkage int sys_mlock(unsigned long start, size_t len)
+#include <asm/uaccess.h>
+#include <asm/pgtable.h>
+
+asmlinkage long sys_mlock(unsigned long start, size_t len)
 {
 	return -ENOSYS;
 }
 
-asmlinkage int sys_munlock(unsigned long start, size_t len)
+asmlinkage long sys_munlock(unsigned long start, size_t len)
 {
 	return -ENOSYS;
 }
 
-asmlinkage int sys_mlockall(int flags)
+asmlinkage long sys_mlockall(int flags)
 {
 	return -ENOSYS;
 }
 
-asmlinkage int sys_munlockall(void)
+asmlinkage long sys_munlockall(void)
 {
 	return -ENOSYS;
 }
